+++ customapi/src/main/java/com/publicissapient/kpidashboard/apis/jira/scrum/service/release/ReleaseBurnUpServiceImpl.java
@@ -467,7 +467,7 @@ public class ReleaseBurnUpServiceImpl extends JiraReleaseKPIService {
 				predictionDataMap.put(ISSUE_SIZE_PREDICTION, issueSizePrediction);
 				boolean isPredictionBoundary = true; // used to check and add data point of release
 				// scope,progress,prediction at boundary
-				for (int i = 0; i <= range && !startLocalDate.isAfter(predictionEndDate); i++) {
+				for (long i = 0; i <= range && !startLocalDate.isAfter(predictionEndDate); i++) {
 					DataCountGroup issueCount = new DataCountGroup();
 					DataCountGroup issueSize = new DataCountGroup();
 					CustomDateRange dateRange = KpiDataHelper.getStartAndEndDateForDataFiltering(startLocalDate,
